Refactor handleValidation
